Don't add _patch_method in base model
We don't need it to be available locally.
