+++ statechart/models/__init__.py
@@ -1,3 +1,2 @@
 from . import statechart
 from . import statechart_mixin
-from . import base
